+++ src/components/Trip/TripProvider.jsx
@@ -1,0 +1,21 @@
+import React, { useMemo, useContext } from 'react'
+
+export const TripContext = React.createContext()
+
+export const useTrip = () => {
+  return useContext(TripContext)
+}
+
+const TripProvider = ({ geojson, trip, children }) => {
+  const value = useMemo(
+    () => ({
+      geojson,
+      trip
+    }),
+    [geojson, trip]
+  )
+
+  return <TripContext.Provider value={value}>{children}</TripContext.Provider>
+}
+
+export default React.memo(TripProvider)