+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add view with specific url to show a trip
 * Add a bottom sheet to show trip information on the trip view
+* Add mode edition to change a mode for a section in the trip view
 
 ## 🐛 Bug Fixes